Fixed the normalized misfit handling in waveform plots.
Changed default behavior for collect_synthetics (but also collect_attributes, even if not really used much in the code -- for consistency).

I also noticed that if the normalized misfit per trace was not stored as attributes in the level0, then it would divide the raw misfit by the total (normalized) misfit. So in the 2009 eq event, this translates to doing ~1e-9 / ~1e0

'hybrid' had not been properly implemented, so I did a light rework of the calculate_norm_data function.